Move SqlTable and TimeSpineSource to metricflow-semantics
Needed to allow using TimeSpineSouce in LinkableSpecResolver
