+++ tests_metricflow/plan_conversion/test_time_spine.py
@@ -5,8 +5,8 @@ import textwrap
 from dbt_semantic_interfaces.type_enums import TimeGranularity
 from metricflow_semantics.filters.time_constraint import TimeRangeConstraint
 from metricflow_semantics.model.semantic_manifest_lookup import SemanticManifestLookup
+from metricflow_semantics.time.time_spine_source import TimeSpineSource
 
-from metricflow.plan_conversion.time_spine import TimeSpineSource
 from metricflow.protocols.sql_client import SqlClient
