+++ tests_metricflow/sql/optimizer/test_table_alias_simplifier.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 import pytest
 from _pytest.fixtures import FixtureRequest
 from metricflow_semantics.sql.sql_join_type import SqlJoinType
+from metricflow_semantics.sql.sql_table import SqlTable
 from metricflow_semantics.test_helpers.config_helpers import MetricFlowTestConfiguration
 
 from metricflow.sql.optimizer.table_alias_simplifier import SqlTableAliasSimplifier
@@ -19,7 +20,6 @@ from metricflow.sql.sql_plan import (
     SqlSelectStatementNode,
     SqlTableFromClauseNode,
 )
-from metricflow.sql.sql_table import SqlTable
 from tests_metricflow.sql.compare_sql_plan import assert_default_rendered_sql_equal
