+++ tests_metricflow/sql/test_sql_plan_render.py
@@ -6,6 +6,7 @@ from typing import List
 import pytest
 from _pytest.fixtures import FixtureRequest
 from metricflow_semantics.sql.sql_join_type import SqlJoinType
+from metricflow_semantics.sql.sql_table import SqlTable, SqlTableType
 from metricflow_semantics.test_helpers.config_helpers import MetricFlowTestConfiguration
 
 from metricflow.protocols.sql_client import SqlClient
@@ -26,7 +27,6 @@ from metricflow.sql.sql_plan import (
     SqlSelectStatementNode,
     SqlTableFromClauseNode,
 )
-from metricflow.sql.sql_table import SqlTable, SqlTableType
 from tests_metricflow.sql.compare_sql_plan import assert_rendered_sql_equal
 
 logger = logging.getLogger(__name__)
