+++ tests_metricflow/sql_clients/test_sql_client.py
@@ -7,11 +7,11 @@ import pytest
 from dbt_semantic_interfaces.test_utils import as_datetime
 from metricflow_semantics.random_id import random_id
 from metricflow_semantics.sql.sql_bind_parameters import SqlBindParameters
+from metricflow_semantics.sql.sql_table import SqlTable
 from metricflow_semantics.test_helpers.config_helpers import MetricFlowTestConfiguration
 
 from metricflow.data_table.mf_table import MetricFlowDataTable
 from metricflow.protocols.sql_client import SqlClient, SqlEngine
-from metricflow.sql.sql_table import SqlTable
 from tests_metricflow.fixtures.sql_clients.ddl_sql_client import SqlClientWithDDLMethods
 from tests_metricflow.sql.compare_data_table import assert_data_tables_equal
